Introduce non-blocking ContainerRuntime
