+++ utils/container-runtime/Cargo.toml
@@ -16,4 +16,6 @@ dill = "0.5"
 regex = "1"
 serde = { version = "1", features = ["derive"] }
 thiserror = "1"
+tokio = { version = "1", default-features = false, features=["process", "sync", "parking_lot"] }
+tracing = "0.1"
 url = "2"
