+++ utils/container-runtime/src/lib.rs
@@ -11,5 +11,16 @@
 #![feature(provide_any)]
 #![feature(error_generic_member_access)]
 
-mod container_runtime;
-pub use crate::container_runtime::*;
+mod args;
+mod blocking;
+mod common;
+mod config;
+mod errors;
+mod listener;
+pub mod nonblocking;
+
+pub use args::*;
+pub use blocking::*; // TODO: Deprecate
+pub use config::*;
+pub use errors::{ContainerRuntimeError, ImagePullError, TimeoutError};
+pub use listener::*;
