+++ utils/container-runtime/src/listener.rs
@@ -1,0 +1,16 @@
+// Copyright Kamu Data, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+pub trait PullImageListener {
+    fn begin(&self, _image: &str) {}
+    fn success(&self) {}
+}
+
+pub struct NullPullImageListener;
+impl PullImageListener for NullPullImageListener {}
